requirements.txt
asgiref==3.6.0
Django==4.1.4
sqlparse==0.4.3
numpy==1.25.1

psycopg2==2.9.3
uvicorn[standard]==0.20.0
whitenoise==6.3.0
django-cors-headers==3.13.0

django-environ==0.9.0
django-cleanup==6.0.0
django-environ==0.9.0
django-extensions==3.2.0
djangorestframework==3.13.1
django-leaflet==0.28.3
django-shortuuidfield==0.1.3
django-fieldsignals==0.7.0
django-silk==5.0.3
django-phonenumbers==1.0.1

PyJWT==2.6.0
httpx[http2]==0.23.3

twilio==8.0.0

celery==5.2.7
django-celery-beat==2.4.0

google==3.0.0
google-auth==2.11.0
shortuuid==1.0.1
email-validator==1.3.0
requests==2.28.0
gdal==3.7.1.1
haversine==2.7.0
Pillow==9.3.0

channels==3.0.5
channels-redis==3.4.1
redis==4.4.0
qrcode==7.3
pydash==6.0.2
python-dotenv==1.0.0
django-concurrency==2.4

boto3==1.26.59
django-storages==1.14